+++ pytest_container/runtime.py
@@ -23,8 +23,6 @@ import pytest
 import testinfra
 from _pytest.mark.structures import ParameterSet
 
-from pytest_container.inspect import _DockerImageInspect
-from pytest_container.inspect import _PodmanImageInspect
 from pytest_container.inspect import BindMount
 from pytest_container.inspect import Config
 from pytest_container.inspect import ContainerHealth
@@ -217,15 +215,6 @@ class OciRuntimeABC(ABC):
     @abstractmethod
     def version(self) -> Version:
         """The version of the container runtime."""
-
-    @abstractmethod
-    def get_container_healthcheck(
-        self, container_image: Union[str, "ContainerBase"]
-    ) -> Optional[HealthCheck]:
-        """Obtain the container image's ``HEALTCHECK`` if defined by the
-        container image or ``None`` otherwise.
-
-        """
 
     @abstractmethod
     def inspect_container(self, container_id: str) -> ContainerInspect:
@@ -463,23 +452,6 @@ class PodmanRuntime(OciRuntimeBase):
             LOCALHOST.run_expect([0], "podman --version").stdout
         )
 
-    def get_container_healthcheck(
-        self, container_image: Union[str, "ContainerBase"]
-    ) -> Optional[HealthCheck]:
-        img_inspect_list: List[_PodmanImageInspect] = json.loads(
-            LOCALHOST.run_expect(
-                [0], f"podman inspect {container_image}"
-            ).stdout
-        )
-        if len(img_inspect_list) != 1:
-            raise RuntimeError(
-                f"Inspecting {container_image} resulted in {len(img_inspect_list)} images"
-            )
-        img_inspect = img_inspect_list[0]
-        if "Healthcheck" not in img_inspect:
-            return None
-        return HealthCheck.from_container_inspect(img_inspect["Healthcheck"])
-
     def inspect_container(self, container_id: str) -> ContainerInspect:
         inspect = self._get_container_inspect(container_id)
 
@@ -562,27 +534,6 @@ class DockerRuntime(OciRuntimeBase):
         """Returns the version of docker installed on this system"""
         return _get_docker_version(
             LOCALHOST.run_expect([0], "docker --version").stdout
-        )
-
-    def get_container_healthcheck(
-        self, container_image: Union[str, "ContainerBase"]
-    ) -> Optional[HealthCheck]:
-        img_inspect_list: List[_DockerImageInspect] = json.loads(
-            LOCALHOST.run_expect(
-                [0], f"docker inspect {str(container_image)}"
-            ).stdout
-        )
-        if len(img_inspect_list) != 1:
-            raise RuntimeError(
-                f"Inspecting {container_image} resulted in {len(img_inspect_list)} images"
-            )
-        img_inspect = img_inspect_list[0]
-        if "Config" not in img_inspect:
-            return None
-        if "Healthcheck" not in img_inspect["Config"]:
-            return None
-        return HealthCheck.from_container_inspect(
-            img_inspect["Config"]["Healthcheck"]
         )
 
     def inspect_container(self, container_id: str) -> ContainerInspect: